style: cos模块重命名
